CON-359: Split up code related to VotingMatrix.
