+++ casper/src/test/scala/io/casperlabs/casper/finality/votingmatrix/VotingMatrixTest.scala
@@ -1,20 +1,25 @@
 package io.casperlabs.casper.finality.votingmatrix
 
+import cats.Monad
 import cats.mtl.MonadState
 import com.github.ghik.silencer.silent
+import cats.implicits._
 import com.google.protobuf.ByteString
+import io.casperlabs.blockstorage.{BlockMetadata, BlockStorage, IndexedDagStorage}
 import io.casperlabs.casper.Estimator.{BlockHash, Validator}
-import io.casperlabs.casper.consensus.Bond
-import io.casperlabs.casper.finality.CommitteeWithConsensusValue
+import io.casperlabs.casper.consensus.{Block, Bond}
 import io.casperlabs.casper.finality.votingmatrix.FinalityDetectorVotingMatrix._votingMatrixS
-import io.casperlabs.casper.finality.votingmatrix.VotingMatrixImpl.VotingMatrixState
+import io.casperlabs.casper.finality.votingmatrix.VotingMatrix.VotingMatrix
+import io.casperlabs.casper.finality.{CommitteeWithConsensusValue, FinalityDetectorUtil}
 import io.casperlabs.casper.helper.BlockUtil.generateValidator
 import io.casperlabs.casper.helper.{BlockGenerator, DagStorageFixture}
+import io.casperlabs.casper.util.ProtoUtil
 import io.casperlabs.p2p.EffectsTestInstances.LogStub
+import io.casperlabs.shared.Time
 import monix.eval.Task
 import org.scalatest.{Assertion, FlatSpec, Matchers}
 
-import scala.collection.immutable.Map
+import scala.collection.immutable.{HashMap, Map}
 
 @silent("is never used")
 class VotingMatrixTest extends FlatSpec with Matchers with BlockGenerator with DagStorageFixture {
@@ -64,7 +69,7 @@ class VotingMatrixTest extends FlatSpec with Matchers with BlockGenerator with D
   )(implicit _votingMatrixS: _votingMatrixS[Task]): Task[Assertion] =
     for {
       votingMatrixState <- MonadState[Task, VotingMatrixState].get
-      result = votingMatrixState.firstLevelZeroVotes shouldBe (VotingMatrixImpl.fromMapToArray(
+      result = votingMatrixState.firstLevelZeroVotes shouldBe (FinalityDetectorUtil.fromMapToArray(
         votingMatrixState.validatorToIdx,
         expect
       ))
@@ -119,7 +124,7 @@ class VotingMatrixTest extends FlatSpec with Matchers with BlockGenerator with D
                   v2 -> None
                 )
               )
-          committee <- VotingMatrix.checkForCommittee[Task](0.1)
+          committee <- checkForCommittee[Task](0.1)
           _         = committee shouldBe None
           b2 <- createAndUpdateVotingMatrix[Task](
                  Seq(genesis.blockHash),
@@ -139,7 +144,7 @@ class VotingMatrixTest extends FlatSpec with Matchers with BlockGenerator with D
                   v2 -> Some((b2.blockHash, b2.getHeader.rank))
                 )
               )
-          committee <- VotingMatrix.checkForCommittee[Task](0.1)
+          committee <- checkForCommittee[Task](0.1)
           _         = committee shouldBe None
           b3 <- createAndUpdateVotingMatrix[Task](
                  Seq(b1.blockHash),
@@ -160,7 +165,7 @@ class VotingMatrixTest extends FlatSpec with Matchers with BlockGenerator with D
                   v2 -> Some((b2.blockHash, b2.getHeader.rank))
                 )
               )
-          committee <- VotingMatrix.checkForCommittee[Task](0.1)
+          committee <- checkForCommittee[Task](0.1)
           _         = committee shouldBe None
           b4 <- createAndUpdateVotingMatrix[Task](
                  Seq(b3.blockHash),
@@ -182,7 +187,7 @@ class VotingMatrixTest extends FlatSpec with Matchers with BlockGenerator with D
                 )
               )
 
-          committee <- VotingMatrix.checkForCommittee[Task](0.01)
+          committee <- checkForCommittee[Task](0.01)
           _         = committee shouldBe None
 
           b5 <- createAndUpdateVotingMatrix[Task](
@@ -205,10 +210,10 @@ class VotingMatrixTest extends FlatSpec with Matchers with BlockGenerator with D
                 )
               )
 
-          committee <- VotingMatrix.checkForCommittee[Task](0.1)
+          committee <- checkForCommittee[Task](0.1)
           _         = committee shouldBe Some(CommitteeWithConsensusValue(Set(v1, v2), 20, b1.blockHash))
 
-          committee <- VotingMatrix.checkForCommittee[Task](0.4)
+          committee <- checkForCommittee[Task](0.4)
           _ = committee shouldBe Some(
             CommitteeWithConsensusValue(Set(v1, v2), 20, b1.blockHash)
           )
@@ -236,4 +241,23 @@ class VotingMatrixTest extends FlatSpec with Matchers with BlockGenerator with D
         } yield result
   }
 
+  def createAndUpdateVotingMatrix[F[_]: Monad: Time: BlockStorage: IndexedDagStorage](
+      parentsHashList: Seq[BlockHash],
+      latestFinalizedBlockHash: BlockHash,
+      creator: Validator = ByteString.EMPTY,
+      bonds: Seq[Bond] = Seq.empty[Bond],
+      justifications: collection.Map[Validator, BlockHash] = HashMap.empty[Validator, BlockHash]
+  )(
+      implicit votingMatrix: VotingMatrix[F]
+  ): F[Block] =
+    for {
+      b           <- createBlock[F](parentsHashList, creator, bonds, justifications)
+      dag         <- IndexedDagStorage[F].getRepresentation
+      votedBranch <- ProtoUtil.votedBranch(dag, latestFinalizedBlockHash, b.blockHash)
+      _ <- updateVoterPerspective(
+            dag,
+            BlockMetadata.fromBlock(b),
+            votedBranch.get
+          )
+    } yield b
 }
